Pen Test

+++ src/routes/userRouter.js
@@ -77,10 +77,9 @@ userRouter.put(
     const { name, email, password } = req.body;
     const userId = Number(req.params.userId);
     const user = req.user;
-    if (user.id !== userId && !user.isRole(Role.Admin)) {
+    if (user.id !== userId) {
       return res.status(403).json({ message: 'unauthorized' });
     }
-
     const updatedUser = await DB.updateUser(userId, name, email, password);
     const auth = await setAuth(updatedUser);
     res.json({ user: updatedUser, token: auth });
@@ -96,7 +95,10 @@ userRouter.delete(
   logger.httpLogger,
   authRouter.authenticateToken,
   asyncHandler(async (req, res) => {
-    const userId = Number(req.params.userId)
+    const userId = Number(req.params.userId);
+    if (!user.isRole(Role.Admin)) {
+      return res.status(403).json({ message: 'unauthorized' });
+    }
     await DB.deleteUser(userId);
     res.locals.auth = req.headers.authorization.split(' ')[1];
     res.json({ message: 'user deleted' });
